requirements.txt
fastapi==0.111.0
uvicorn==0.30.0
#motor==3.5.1
pymongo==4.8.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic[email]==2.7.4
python-dotenv==1.0.1
